Added: install script

+++ update.sh
@@ -1,17 +1,18 @@
 #!/usr/bin/env bash
 
-source inc/functions.sh
-init_ansi
-
-print_header "Copying dotfiles"
+echo "Copying dotfiles"
 
 dotfiles=(".bash_profile" ".gitconfig" ".inputrc" ".tmux.conf" ".vimrc" ".zshrc")
 
 for f in ${dotfiles[@]}; do
-  print_bullet ${f}
+  if [[ -L ~/{f} ]]; then
+   echo "Skipping symlink for ${f}"
+   continue 
+  fi
+  echo "${f}"
   cp "${HOME}/${f}" .
 done
 
-print_bullet 🐟
+echo "🐟"
 cp -r ${HOME}/.config/fish .
